+++ src/InstructionMemory.java
@@ -1,11 +1,11 @@
 public class InstructionMemory {
-    private static short[] memory = new short[1024]; //each row id 16 bits (2 bytes)
+    private static int[] memory = new int[1024]; //each row id 16 bits (2 bytes)
 
-    public static short[] getMemory() {
+    public static int[] getMemory() {
         return memory;
     }
 
-    public static void setMemory(short[] memory) {
+    public static void setMemory(int[] memory) {
         InstructionMemory.memory = memory;
     }
 }
